add auth persistant state
