+++ app/wishlists/page.tsx
@@ -1,0 +1,11 @@
+"use client";
+import withAuth from '../../hoc/withAuth'
+
+const WishlistPage = () => {
+
+  return (
+<></>
+  );
+};
+
+export default withAuth(WishlistPage);
