+++ package.json
@@ -10,6 +10,7 @@
   },
   "dependencies": {
     "axios": "^1.7.2",
+    "cookie": "^0.6.0",
     "next": "14.1.3",
     "next-themes": "^0.2.1",
     "react": "^18",
